requirements.txt
django
djangorestframework
django-cors-headers
django-bootstrap5
django-crispy-forms
celery
psycopg2-binary
celery[redis]

